Update: TBB 2.2 20090908 (development release).

+++ examples/concurrent_hash_map/count_strings/count_strings.cpp
@@ -30,7 +30,6 @@
 #if __linux__ && defined(__INTEL_COMPILER)
 #define __sync_fetch_and_add(ptr,addend) _InterlockedExchangeAdd(const_cast<void*>(reinterpret_cast<volatile void*>(ptr)), addend)
 #endif
-
 #include <string>
 #include <cstring>
 #include <cctype>
@@ -59,7 +58,7 @@ static bool Verbose = false;
 static int NThread = 1;
 
 //! Problem size
-size_t N = 1000000;
+long N = 1000000;
 const int size_factor = 2;
 
 //! Indicates if the number of threads wasn't set explicitly
